Package name to lower case
It's a Java convention to avoid conflict with
the names of classes or interfaces.
