+++ src/main/java/com/brianmcmichael/sagu/ContextMenuMouseListener.java
@@ -16,7 +16,7 @@
 //    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 //////////////////////////////////////////////////////////////////////////////////
 
-package com.brianmcmichael.SimpleGlacierUploader;
+package com.brianmcmichael.sagu;
 
 import javax.swing.*;
 import javax.swing.text.JTextComponent;
