+++ src/main/java/com/brianmcmichael/sagu/InventoryRequest.java
@@ -16,7 +16,7 @@
 //    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 //////////////////////////////////////////////////////////////////////////////////
 
-package com.brianmcmichael.SimpleGlacierUploader;
+package com.brianmcmichael.sagu;
 
 import com.amazonaws.AmazonClientException;
 import com.amazonaws.AmazonServiceException;
